fix unsigned short

+++ src/com/descabouda/model/BaseAttribute.scala
@@ -1,7 +1,7 @@
 package com.descabouda.model
 
 class BaseAttribute {
-  var attribute_name_index: Short = null
+  var attribute_name_index: Integer = null
   var attribute_length: Integer = null
   var info: Array[Byte] = null
 }
