+++ src/com/descabouda/model/BaseMethod.scala
@@ -3,10 +3,10 @@ package com.descabouda.model
 import java.util
 
 class BaseMethod {
-  var access_flags: Short = null
-  var name_index: Short = null
-  var descriptor_index: Short = null
-  var attributes_count: Short = null
+  var access_flags: Integer = null
+  var name_index: Integer = null
+  var descriptor_index: Integer = null
+  var attributes_count: Integer = null
 
   val attributes = new util.LinkedList[BaseAttribute]()
 }
