+++ src/com/descabouda/model/constants/MethodTypeConstant.scala
@@ -3,5 +3,5 @@ package com.descabouda.model.constants
 import com.descabouda.model.BaseConstant
 
 class MethodTypeConstant extends BaseConstant {
-  var descriptor_index: Short = null
+  var descriptor_index: Integer = null
 }
