+++ src/com/descabouda/model/constants/NameAndTypeConstant.scala
@@ -3,6 +3,6 @@ package com.descabouda.model.constants
 import com.descabouda.model.BaseConstant
 
 class NameAndTypeConstant extends BaseConstant {
-  var name_index: Short = null
-  var descriptor_index: Short = null
+  var name_index: Integer = null
+  var descriptor_index: Integer = null
 }
